# ApnaStayBackEnd/docs/swagger.json
  {
    "openapi": "3.0.3",
    "info": {
      "title": "HomeFinity Backend API",
      "description": "API for HomeFinity frontend.\n\n**Required headers for protected endpoints:**\n- **Authorization** (mandatory): `Authorization: Bearer <JWT>` — JWT from POST /api/auth/public/signin or POST /api/auth/public/phone/verify-and-login.\n- **X-XSRF-TOKEN** (mandatory): CSRF token from GET /api/csrf-token.\n\n**Required header for requests with body (POST/PUT):**\n- **Content-Type** (mandatory): `Content-Type: application/json`.\n\nPublic endpoints (signin, signup, phone send-code, phone verify-and-login, forgot-password, reset-password, verify-2fa-login, csrf-token, health-check) do not require Authorization or X-XSRF-TOKEN.",
      "version": "0.2.0"
    },
    "servers": [
      {
        "url": "/",
        "description": "Relative to deployed base URL"
      }
    ],
...
    "security": [
      {
        "bearerAuth": [],
        "csrfHeader": []
      }
    ],
    "paths": {
...
      "/api/auth/username": {
        "get": {
          "tags": ["Auth"],
          "summary": "Get current username",
          "description": "Returns the authenticated user's username. Requires Authorization and X-XSRF-TOKEN (mandatory).",
          "operationId": "getUsername",
"parameters": [
          { "$ref": "#/components/parameters/XsrfTokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Username or error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": { "type": "boolean" },
                      "message": { "type": "string" },
                      "timestamp": { "type": "string", "format": "date-time" },
                      "data": { "type": "string", "description": "Username when success is true" }
                    }
                  }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
...
      "/api/complaints": {
        "get": {
          "tags": ["Complaint"],
          "summary": "List complaints",
          "description": "Returns complaints for the current user (filtered by role). Admin: all; Owner/User: raised by, assigned to, or related. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "listComplaints",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "status", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/ComplaintStatus" } }
          ],
          "responses": {
            "200": {
              "description": "List of complaints",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseComplaintList" }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        },
        "post": {
          "tags": ["Complaint"],
          "summary": "Raise complaint",
          "description": "Create a new complaint. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "raiseComplaint",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComplaintRequest" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Complaint raised successfully",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseComplaint" }
                }
              }
            },
            "400": { "description": "Validation error" },
            "401": { "description": "Unauthorized" }
          }
        }
      },
      "/api/complaints/{id}": {
        "get": {
          "tags": ["Complaint"],
          "summary": "Get complaint by ID",
          "description": "Get a complaint by ID. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "getComplaint",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
          ],
          "responses": {
            "200": {
              "description": "Complaint details",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseComplaint" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "404": { "description": "Complaint not found" }
          }
        }
      },
      "/api/complaints/{id}/resolve": {
        "put": {
          "tags": ["Complaint"],
          "summary": "Resolve complaint",
          "description": "Mark a complaint as resolved. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "resolveComplaint",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
          ],
          "requestBody": {
            "required": false,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResolveComplaintRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Complaint resolved",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseComplaint" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" },
            "404": { "description": "Complaint not found" }
          }
        }
      },
      "/api/complaints/{id}/assign": {
        "put": {
          "tags": ["Complaint"],
          "summary": "Assign complaint",
          "description": "Assign a complaint to a user. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "assignComplaint",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AssignComplaintRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Complaint assigned",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseComplaint" }
                }
              }
            },
            "400": { "description": "Validation error" },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" },
            "404": { "description": "Complaint not found" }
          }
        }
      },
      "/api/complaints/{id}/status": {
        "put": {
          "tags": ["Complaint"],
          "summary": "Update complaint status",
          "description": "Update complaint status. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "updateComplaintStatus",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
            { "name": "status", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/ComplaintStatus" } }
          ],
          "responses": {
            "200": {
              "description": "Status updated",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseComplaint" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" },
            "404": { "description": "Complaint not found" }
          }
        }
      },
      "/api/complaints/{id}/messages": {
        "get": {
          "tags": ["Complaint"],
          "summary": "Get complaint messages",
          "description": "Get all messages for a complaint. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "getComplaintMessages",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
          ],
          "responses": {
            "200": {
              "description": "List of messages",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseComplaintMessageList" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "404": { "description": "Complaint not found" }
          }
        },
        "post": {
          "tags": ["Complaint"],
          "summary": "Add message to complaint",
          "description": "Add a message to a complaint. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "addComplaintMessage",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComplaintMessageRequest" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Message sent",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseComplaintMessage" }
                }
              }
            },
            "400": { "description": "Validation error" },
            "401": { "description": "Unauthorized" },
            "404": { "description": "Complaint not found" }
          }
        }
      },
...
    },
    "components": {
      "parameters": {
        "XsrfTokenHeader": {
          "in": "header",
          "name": "X-XSRF-TOKEN",
          "required": true,
          "schema": {
            "type": "string",
            "description": "CSRF token from GET /api/csrf-token (use the token value from response)"
          },
          "description": "**Mandatory** for all protected endpoints. Obtain token from GET /api/csrf-token. Authorization is defined via securitySchemes (Bearer JWT). Content-Type is defined by requestBody.content for operations with a body."
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "JWT token from /api/auth/public/signin. Send as: Authorization: Bearer <token>"
        },
        "csrfHeader": {
          "type": "apiKey",
          "in": "header",
          "name": "X-XSRF-TOKEN",
          "description": "CSRF token from GET /api/csrf-token. Use the header name returned by that endpoint (e.g. X-XSRF-TOKEN) and the token value. Required for state-changing requests and optionally for GETs when cookies are used."
        }
      },
      "schemas": {
...
        "ComplaintStatus": {
          "type": "string",
          "enum": ["OPEN", "IN_PROGRESS", "RESOLVED", "CLOSED"]
        },
        "ComplaintPriority": {
          "type": "string",
          "enum": ["LOW", "MEDIUM", "HIGH"]
        },
        "ComplaintRequest": {
          "type": "object",
          "required": ["subject", "description"],
          "properties": {
            "subject": { "type": "string", "maxLength": 255 },
            "description": { "type": "string", "maxLength": 5000 },
            "priority": { "$ref": "#/components/schemas/ComplaintPriority" },
            "relatedUserId": { "type": "integer", "format": "int64" },
            "propertyId": { "type": "integer", "format": "int64" }
          }
        },
        "ComplaintDTO": {
          "type": "object",
          "properties": {
            "id": { "type": "integer", "format": "int64" },
            "raisedByUserId": { "type": "integer", "format": "int64" },
            "raisedByUserName": { "type": "string" },
            "assignedToUserId": { "type": "integer", "format": "int64" },
            "assignedToUserName": { "type": "string" },
            "relatedUserId": { "type": "integer", "format": "int64" },
            "relatedUserName": { "type": "string" },
            "propertyId": { "type": "integer", "format": "int64" },
            "subject": { "type": "string" },
            "description": { "type": "string" },
            "status": { "$ref": "#/components/schemas/ComplaintStatus" },
            "priority": { "$ref": "#/components/schemas/ComplaintPriority" },
            "resolutionNote": { "type": "string" },
            "resolvedAt": { "type": "string", "format": "date-time" },
            "resolvedByUserId": { "type": "integer", "format": "int64" },
            "resolvedByUserName": { "type": "string" },
            "createdAt": { "type": "string", "format": "date-time" },
            "updatedAt": { "type": "string", "format": "date-time" },
            "messages": { "type": "array", "items": { "$ref": "#/components/schemas/ComplaintMessageDTO" } }
          }
        },
        "ResolveComplaintRequest": {
          "type": "object",
          "properties": {
            "resolutionNote": { "type": "string" }
          }
        },
        "AssignComplaintRequest": {
          "type": "object",
          "required": ["assignToUserId"],
          "properties": {
            "assignToUserId": { "type": "integer", "format": "int64" }
          }
        },
        "ComplaintMessageRequest": {
          "type": "object",
          "required": ["messageText"],
          "properties": {
            "messageText": { "type": "string", "maxLength": 2000 }
          }
        },
        "ComplaintMessageDTO": {
          "type": "object",
          "properties": {
            "id": { "type": "integer", "format": "int64" },
            "complaintId": { "type": "integer", "format": "int64" },
            "senderId": { "type": "integer", "format": "int64" },
            "senderUserName": { "type": "string" },
            "messageText": { "type": "string" },
            "createdAt": { "type": "string", "format": "date-time" }
          }
        },
        "ApiResponseComplaint": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "$ref": "#/components/schemas/ComplaintDTO" }
          }
        },
        "ApiResponseComplaintList": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "type": "array", "items": { "$ref": "#/components/schemas/ComplaintDTO" } }
          }
        },
        "ApiResponseComplaintMessage": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "$ref": "#/components/schemas/ComplaintMessageDTO" }
          }
        },
        "ApiResponseComplaintMessageList": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "type": "array", "items": { "$ref": "#/components/schemas/ComplaintMessageDTO" } }
          }
        },
...
      }
    }
  }