# ApnaStayBackEnd/src/main/java/com/secure/apnastaybackend/dto/request/ResolveComplaintRequest.java
package com.secure.apnastaybackend.dto.request;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ResolveComplaintRequest {

    private String resolutionNote;
}