# ApnaStayBackEnd/src/main/java/com/secure/apnastaybackend/dto/response/ComplaintMessageDTO.java
package com.secure.apnastaybackend.dto.response;

import com.fasterxml.jackson.annotation.JsonFormat;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ComplaintMessageDTO {

    private Long id;
    private Long complaintId;
    private Long senderId;
    private String senderUserName;
    private String messageText;

    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
    private LocalDateTime createdAt;
}